реализована полностью выдача и фильтр по выдаче

+++ src/main/java/ru/vp/library/dto/BookDTO.java
@@ -31,6 +31,6 @@ public class BookDTO {
     private int pageCount;
     private String location;
     private int price;
-    private int totalNumber;
+    private long totalNumber;
     private long totalAvailableNumber;
 }
